+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use `app-platform` chart to bootstrap CRDs and custom resources.
+
 ## [0.10.0] - 2021-08-09
 
 ### Added